Move auth-client from repository to client
